requirements.txt
Flask
geocoder
python-dotenv
pygame
speechrecognition
